+++ src/commonTest/kotlin/csense/kotlin/extensions/primitives/StringTest.kt
@@ -77,16 +77,16 @@ class StringTest {
 
     @Test
     fun findAllOf() {
-        "".findAllOf("", searchByWord = false, ignoreCase = false)
+        "".allIndicesOf("", searchByWord = false, ignoreCase = false)
             .assertEmpty("nothing in nothing is nothing")
         val textString = "\"a very funny qoute\""
-        textString.findAllOf("\"", searchByWord = false, ignoreCase = false).apply {
+        textString.allIndicesOf("\"", searchByWord = false, ignoreCase = false).apply {
             assertSize(2, "since there are 2 \" in the text")
             first().assert(0, "since the first \" is at the first location")
             last().assert(textString.length - 1, "since the last \" is at the end")
         }
 
-        textString.findAllOf("abc", searchByWord = false, ignoreCase = false).assertEmpty()
+        textString.allIndicesOf("abc", searchByWord = false, ignoreCase = false).assertEmpty()
     }
 
     @Test
